# api/openapi.yaml
openapi: 3.0.0
info:
  title: HR Recruiter API
  version: 1.0.0
  description: |
    Полная OpenAPI спецификация для платформы автоматизации.
    Включает все основные сущности, бизнес-операции, заглушки для AI и медиа, а также примеры для быстрой интеграции.
servers:
  - url: http://localhost:8080/api/v1
    description: Local server
  - url: https://api.hr-recruiter.com/api/v1
    description: Production server
...
paths:
...
  /account:
    get:
      operationId: getAccount
      tags:
        - Account
      summary: Получить профиль текущего пользователя
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      operationId: updateAccount
      tags:
        - Account
      summary: Обновить профиль текущего пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Профиль обновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /ai/transcribe:
    post:
      operationId: transcribeAudio
      tags:
        - AI
      summary: Транскрибировать аудио файл
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audio
              properties:
                audio:
                  type: string
                  format: binary
                  description: Аудио файл для транскрибации
      responses:
        '200':
          description: Аудио успешно транскрибировано
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcript:
                    type: string
                    description: Транскрибированный текст
  /ai/generate-questions:
    post:
      operationId: generateQuestions
      tags:
        - AI
      summary: Генерировать вопросы для собеседования
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateQuestionsRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
        '500':
          description: Internal Server Error
  /ai/generate-position:
    post:
      operationId: generatePosition
      tags:
        - AI
      summary: Генерировать вакансию с помощью AI
      description: |
        Генерирует структуру вакансии на основе описания пользователя.
        AI возвращает минимальный набор полей (title, description, topics, level, questions).
        Остальные поля (status, language, settings и т.д.) устанавливаются системой по умолчанию.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionAiGenerationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionAiGenerationResponse'
        '400':
          description: Bad Request - некорректные входные данные
        '500':
          description: Internal Server Error - ошибка AI или системы
...
  /ai/generate-position-data:
    post:
      operationId: generatePositionData
      tags:
        - AI
      summary: Сгенерировать данные для вакансии на основе описания
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PositionDataGenerationRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDataGenerationResponse'
        '400':
          description: Неверный запрос
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Ошибка генерации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /ai/transcribe-answer:
    post:
      operationId: transcribeAnswerWithAI
      tags:
        - AI
      summary: Транскрибировать и отформатировать аудио ответ на интервью
      description: |
        Обрабатывает аудио файл через пайплайн:
        1. Whisper - транскрибация в сырой текст
        2. Claude - форматирование текста (без добавления контента)
        3. Сохранение результата в базу данных

        Сохраняет как сырой, так и отформатированный текст.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audioFile
                - interviewId
                - questionId
              properties:
                audioFile:
                  type: string
                  format: binary
                  description: Аудио файл для транскрибации
                interviewId:
                  type: integer
                  format: int64
                  description: ID интервью (определяет кандидата и позицию)
                questionId:
                  type: integer
                  format: int64
                  description: ID вопроса для создания ответа
      responses:
        '200':
          description: Аудио успешно обработано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeAnswerWithAI200Response'
        '400':
          description: Неверные параметры запроса
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Audio file is empty"
        '404':
          description: Интервью или вопрос не найден
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Interview or question not found"
        '503':
          description: Сервисы транскрибации недоступны
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Transcription services are not available"
...
components:
  schemas:
...
    BaseEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, createdAt]
    
    # Общий класс для пагинированных ответов
...
    BaseQuestionFields:
      type: object
      properties:
        text:
          type: string
        type:
          $ref: '#/components/schemas/QuestionTypeEnum'
        order:
          type: integer
        isRequired:
          type: boolean
        evaluationCriteria:
          type: string
          description: "Критерии оценки ответа"
...
    BaseUserFields:
      type: object
      properties:
        firstName:
          type: string
          description: "Имя пользователя"
        lastName:
          type: string
          description: "Фамилия пользователя"
        email:
          type: string
        role:
          $ref: '#/components/schemas/RoleEnum'
        avatarUrl:
          type: string
        language:
          type: string
        phone:
          type: string
    
    # Основные схемы
...
    User:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseUserFields'
      required: [id, firstName, lastName, email, role, createdAt]
...
    UserUpdateRequest:
      $ref: '#/components/schemas/BaseUserFields'
...
    Question:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - $ref: '#/components/schemas/BaseQuestionFields'
        - type: object
          properties:
            positionId:
              type: integer
          required: [id, positionId, text, type, order, createdAt]
...
    QuestionTypeEnum:
      type: string
      enum:
        - text
        - audio
        - video
        - choice
    PositionLevelEnum:
      type: string
      enum:
        - junior
        - middle
        - senior
        - lead
      description: "Уровень позиции"
    RoleEnum:
      type: string
      enum:
        - admin
        - recruiter
        - viewer
    
    # Stats
...
    PositionDataGenerationRequest:
      type: object
      properties:
        description:
          type: string
          description: "Описание вакансии для генерации данных"
          example: "Нужен Java разработчик для разработки микросервисов"
        existingData:
          type: object
          description: "Существующие данные вакансии (опционально)"
          properties:
            title:
              type: string
            topics:
              type: array
              items:
                type: string
            level:
              $ref: '#/components/schemas/PositionLevelEnum'
            tags:
              type: array
              items:
                type: string
      required: [description]
    PositionDataGenerationResponse:
      type: object
      properties:
        generatedData:
          type: object
          properties:
            title:
              type: string
              description: "Сгенерированное название вакансии"
              example: "Java Backend Developer"
            topics:
              type: array
              items:
                type: string
              description: "Сгенерированные темы для собеседования"
              example: ["Java Core", "Spring Framework", "Microservices", "Database Design"]
            level:
              $ref: '#/components/schemas/PositionLevelEnum'
              description: "Предлагаемый уровень позиции"
            tags:
              type: array
              items:
                type: string
              description: "Сгенерированные теги"
              example: ["Java", "Backend", "Spring", "Microservices"]
            suggestedDescription:
              type: string
              description: "Улучшенное описание вакансии"
        confidence:
          type: number
          format: float
          description: "Уверенность в генерации (0-1)"
          example: 0.85
        message:
          type: string
          description: "Сообщение о результатах генерации"
          example: "Данные успешно сгенерированы на основе описания"
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: "Сообщение об ошибке"
          example: "Произошла ошибка при обработке запроса"
        error:
          type: string
          description: "Тип ошибки"
          example: "VALIDATION_ERROR"
        timestamp:
          type: string
          format: date-time
          description: "Время возникновения ошибки"
          example: "2024-01-15T10:30:45Z"
        path:
          type: string
          description: "Путь запроса"
          example: "/api/positions"
      required: [message, error, timestamp]
    TranscribeAnswerWithAI200Response:
      type: object
      properties:
        success:
          type: boolean
          example: true
        formattedText:
          type: string
          description: Отформатированный текст от Claude
          example: "Я работал в компании три года, занимался разработкой на Java."
        interviewAnswerId:
          type: integer
          format: int64
          description: ID ответа на интервью
...
    PositionAiGenerationRequest:
      type: object
      description: "Запрос на генерацию вакансии с помощью AI"
      properties:
        description:
          type: string
          description: "Описание вакансии от пользователя"
          example: "Нужен Java разработчик для разработки микросервисов"
        questionsCount:
          type: integer
          description: "Количество вопросов для собеседования"
          example: 5
          minimum: 1
          maximum: 20
        questionType:
          type: string
          description: "Тип вопросов (hard, soft, mixed)"
          example: "hard"
      required: [description]
    PositionAiGenerationResponse:
      type: object
      description: "Ответ AI с минимальной структурой вакансии"
      properties:
        title:
          type: string
          description: "Название вакансии"
          example: "Java Backend Developer"
        description:
          type: string
          description: "Описание вакансии"
          example: "Разработка микросервисов на Java с использованием Spring Framework"
        topics:
          type: array
          items:
            type: string
          description: "Ключевые темы/навыки для собеседования"
          example: ["Java Core", "Spring Framework", "Microservices", "Database Design"]
        level:
          type: string
          enum: [junior, middle, senior, lead]
          description: "Уровень позиции"
          example: "middle"
        questions:
          type: array
          items:
            $ref: '#/components/schemas/PositionAiQuestion'
          description: "Список вопросов для собеседования"
      required: [title, topics, level, questions]
    PositionAiQuestion:
      type: object
      description: "Вопрос для собеседования (AI-генерация)"
      properties:
        text:
          type: string
          description: "Текст вопроса"
          example: "Расскажите о принципах SOLID"
        type:
          type: string
          enum: [text, audio, choice]
          description: "Тип вопроса"
          example: "text"
        order:
          type: integer
          description: "Порядок вопроса"
          example: 1
      required: [text, type, order]
    GenerateQuestionsRequest:
      type: object
      properties:
        positionDescription:
          type: string
      required: [positionDescription]
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []
...